`Player`: `_iWithinSongs` is now the current index within `_contSongs`.
Previously, the current index used to always be `0` and we used to call `_contSongs.pop_front ()`.
This changeset is a step towards supporting the Repeat and Random features.

`Test.pl` has been updated in order to reflect the newer logic.
When stepping back in `_contHistory` beyond the first item, we simply replay the first item (as before),
but without adding it to `_contSongs`
(since we no longer add from `_contHistory` to `contSongs`)
=> that `Song` does not eventually get re-added to `_contHistory`.