Support for AddSongFromFile. `Lexer`: `ExtractToken` supports filename tokens.
The `ExtractToken` function now has an additional parameter
allowing the user to specify whether a filename is expected,
in which case different lexing rules are applied, e.g.:

* filename not expected:
    `Song1.txt` produces three tokens: `Song1`, `.`, `txt`.

* filename too expected:
    `Song1.txt` produces a single token: `Song1.txt`.

* In either case, quotes can be used to ensure a single token is produced:
    `"Song1.txt"` always produces a single token: `Song1.txt`.

This change has been motivated by the desire to give the user a familiar way
of entering a filename (for the AddSongFromFile command) --
without the need for quotes unless the filename contains whitespace.

In order not to break expectations of existing calls to `ExtractToken`,
this new parameter has a default value of `false`
(=> by default, filename is not expected).